PyModule/requirements.txt
pyserial==3.5
python-dotenv==0.19.2
websocket-client==1.2.3
